Chain the appropriate selectors appropriately
I chained the .avatar and the .proportioned class selectors to
auto adjust the height of the appropriate image in proportion to the
change in width (the image retain its square shape without distortion).
I then chained the .avatar and the .distorted class selectors to
manually hardcode the height of the appropriate image to be twice as
big as its width.