Add descendant combinator for specific text
I added a descendant combinator to the style.css file so that the
elements that contain the text "This should be styled" have a yellow
background, red text, and a font size of 20px. The decscendant
combinator also makes it so that the elements that contain the text
"This should be styled" is center aligned.