Edit style.css to produce the desired outcome
I edited various elements on the style.css file so that
the web page would look identical to the expected outcome for the
exercise. I added margins and padding to the elements. Moreover,
I centered the button-container text and the button. I added "display:
block" under the button selector so that the button does not
sit next to other elements.