Space and align flex items appropriately in header
I used "justify-content: space-between" to appropriately space
the left-links and right-links to the very left and right side of the
header. This also centered the logo on the header horizontally (main
axis). I also used "align-items: center" to make sure that all
of the flex items in the header were centered vertically (cross axis).
I them removed the default margins/padding from the left and right
links. Additionally, I used "gap: 12px" to makse sure that the
links had space between them. The final result was that the desired
outcome was produced for this exercise.