Add containers to produce the desired outcome
I created two containers that were used to appropriately
arrange the flex items. The grouped-items container was used to
center all the items horizontally on the page and space them apart
by 52px. The items containers were used to set the width of each item
to 200px so that the text for them would wrap. The result was that
the web page matched the desired outcome image for this exercise.