Use flexbox to produce the desired alert menu
I created an align-items class to group the header, text, and buttons
together. This class was used to properly place and align
it next to the icon. I also arranged the header class to encompass
the close-button class so that it can be used to appropriately
align and space the desired elements. I also edited the top and bottom
margins of the main text to match the expected outcome.