Center "content" class elements and their spacing
I set the flex-direction of the content to column. Then, I
used "align-items: center" and "justify-content: center" to center
all the elements in the content class. Setting the flex-direction
to column would also ensure that the header and footer remain
at the top and bottom of the page respectively. I also edited the
padding of several elements on the page. Setting the flex-direction
to column in the content class made it so that the logo, input, and
buttons stacked on top of each other vertically.