Position header/footer and "container" elements
I changed the flex-direction to column in the "body" class selector
to push the footer to the bottom of the page. I also added a container
class to the HTML file so that the side bar and the cards would
be lined up horizontally in the middle of the page. I edited the
font size and weight of the header text. Furthermore, I centered the
text in the footer horizontally and vertically.