Edit card section of page to display it properly
I used the card-section class to appropriately dosplay the cards
horizontally. This class was also used to set the margin to 34px and
wrap the cards. I then set the width size of the card elements so that
they would properly wrap when the screen got smaller. Additionally, the
padding and margin of the cards were set.